feat(F-211): integrate with get conversation categories endpoint
