+++ frontend/src/app/[locale]/(main)/new-training/page.tsx
@@ -2,7 +2,7 @@ import { generateMetadata as generateDynamicMetadata } from '@/lib/metadata';
 import type { Metadata } from 'next';
 import { MetadataProps } from '@/interfaces/MetadataProps';
 import { Suspense } from 'react';
-import { getConversationCategories } from '@/services/GetCategories';
+import { getConversationCategories } from '@/services/ConversationCategoriesService';
 import NewTrainingForm from './components/NewTrainingForm';
 import NewTrainingPageLoading from './loading';
